+++ src/pas/plugins/imio/tests/test_migration.py
@@ -46,16 +46,16 @@ class TestMigration(unittest.TestCase):
         self.assertEqual(member.getProperty("email", ""), "")
         self.assertIn("Site Administrator", member.getRoles())
         data = {}
-        data["id"] = "imio"
+        data["id"] = "12345-67890"
         data["username"] = "jamesbond"
         data["email"] = "james@bond.co.uk"
         authomatic_user = User("authentic", **data)
         user = MockupUser(self.plugin, authomatic_user)
         self.plugin.remember_identity(user)
         self.assertNotIn("jamesbond", self.acl_users.source_users.getUserIds())
-        new_user = self.plugin._useridentities_by_userid.get("jamesbond")
-        self.assertEqual(new_user.userid, "jamesbond")
-        member = api.portal.get_tool("portal_membership").getMemberById("jamesbond")
+        new_user = self.plugin._useridentities_by_userid.get("12345-67890")
+        self.assertEqual(new_user.login, "jamesbond")
+        member = api.portal.get_tool("portal_membership").getMemberById("12345-67890")
         self.assertEqual(member.getProperty("email", ""), "james@bond.co.uk")
         self.assertIn("Site Administrator", member.getRoles())