+++ src/pas/plugins/imio/user.py
@@ -21,6 +21,15 @@ class IPASUserDataSchema(IUserDataSchema):
     extra fields.
     """
 
+    # username = schema.TextLine(
+    #     title=_(u"label_user_name", default=u"User Name"),
+    #     description=_(
+    #         u"help_user_name_creation", default=u"Enter user name, e.g. jsmith."
+    #     ),
+    #     required=False,
+    #     readonly=True,
+    # )
+
     fullname = schema.TextLine(
         title=_(u"label_full_name", default=u"Full Name"),
         description=_(